moved function printDACOptions() from dacScanV3.py to qcutilities.py
